USAF aeromedical response to 9/11/01 attacks
This historical study is UNCLASSIFIED in its entirety. It contains no FOUO information. All of the materials mentioned in the bibliography are unclassified. The complete official 932 AW history report for fiscal year 2001 will provide greater details and will include all of the materials mentioned in the bibliography. This historian’s opinions and conclusions do not necessarily reflect those of the United States military. This historical study was reviewed by 932 AW Public Affairs officer 1Lt Stanley E. Paregien, who cleared it for unlimited distribution.