Y2K at USAF's SALC
This UNCLASSIFIED historical paper reveals USAF's Sacramento Air Logistics Center suffered no historically significant issues as a result of the "Y2K" computer issue.  The chronology on p.vii does not acknowledge Y2K as a historically significant event.  A narrative of the Y2K issue begins on p.9 and ends on p.11.