Y2K supporting documents at USAF's SALC
This UNCLASSIFIED historical paper reveals USAF's Sacramento Air Logistics Center suffered no historically significant issues as a result of the "Y2K" computer issue.  Supporting documents' pages are not numbered but you'll find near the front of this PDF a series of pages that lay out a Y2K battle plan that stretches from 0200L on 31 Dec 99 to 0900L on 15 Jan 00.